requirements.txt
docopt==0.6.2
Flask==0.10.1
future==0.15.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.2.1
python-telegram-bot==3.3
tornado==4.3
Werkzeug==0.11.4
